Rename package to trustmodel-local to de-collide with the hosted trustmodel SDK
The open-source engine (evaluate/monitor/govern + MCP server) shipped under the
distribution name 'trustmodel' (import 'trustmodel'), colliding with the official
hosted SDK published to PyPI as 'trustmodel'. Same import name, two different APIs —
customers who pip-installed the SDK then followed docs written against this repo's
API hit AttributeError on Guardrail/Monitor/govern.

- pyproject: name -> trustmodel-local; package -> src/trustmodel_local; scripts point
  at trustmodel_local.* (CLI command names 'trustmodel'/'trustmodel-mcp' unchanged)
- rename import package src/trustmodel -> src/trustmodel_local; update all imports in
  src/examples/tests to 'trustmodel_local'
- README/UPGRADING: install as trustmodel-local; import trustmodel_local; drop the
  'same pip install ships the cloud client' claim — the hosted SDK is a SEPARATE
  package (trustmodel). Two packages, two APIs, no collision.
