+++ UPGRADING.md
@@ -1,73 +1,66 @@
-# Upgrading to TrustModel v3.0.0
+# Upgrading — the local engine is now `trustmodel-local`
 
-v3.0.0 unifies the open-source engine (**Eval / Monitor / Govern**) and the
-**embeddable MCP server** into the same `trustmodel` package that previously
-shipped only the cloud SDK. If you're on **v2.x**, here's how to move up.
+The open-source engine (**Eval / Monitor / Govern**) and the **embeddable MCP server** now ship
+as their own package, **`trustmodel-local`** (import `trustmodel_local`). This is a deliberate
+split from the hosted **`trustmodel`** cloud SDK — two packages, two APIs, no more name collision.
 
 ## TL;DR
 
 ```bash
-pip install -U "trustmodel[mcp]"     # everything: engine + MCP server
+pip install -U "trustmodel-local[mcp]"     # everything: engine + MCP server
 # or, if you don't need the MCP server:
-pip install -U trustmodel            # engine only (evaluate / monitor / govern)
+pip install -U trustmodel-local            # engine only (evaluate / monitor / govern)
 ```
-
-> ⚠️ A plain `pip install trustmodel` will **not** upgrade an existing install —
-> pip treats the requirement as already satisfied and does nothing. The `-U`
-> (`--upgrade`) flag is required.
 
 Verify:
 
 ```bash
-python -c "import trustmodel; print(trustmodel.__version__)"   # → 3.0.0
+python -c "import trustmodel_local; print(trustmodel_local.__version__)"
 ```
 
-## What you get
+## Migrating from the old `trustmodel` engine imports
 
-| Import | New in 3.0.0 | Needs |
-|---|---|---|
-| `from trustmodel import evaluate` | ✅ | base install |
-| `from trustmodel import monitor` | ✅ | base install |
-| `from trustmodel import govern, Guardrail` | ✅ | base install |
-| `trustmodel-mcp` / `trustmodel mcp` (MCP server) | ✅ | `trustmodel[mcp]`, Python ≥ 3.10 |
-| `from trustmodel import TrustModelClient` | unchanged | base install |
+If you were importing the local engine from `trustmodel`, change the module name:
 
-## What stays the same
+| Old | New |
+|---|---|
+| `from trustmodel import evaluate` | `from trustmodel_local import evaluate` |
+| `from trustmodel import monitor` | `from trustmodel_local import monitor` |
+| `from trustmodel import govern, Guardrail` | `from trustmodel_local import govern, Guardrail` |
+| `trustmodel-mcp` / `trustmodel mcp` (MCP server) | unchanged — CLI commands stay `trustmodel` / `trustmodel-mcp` |
 
-- **`TrustModelClient` and all exceptions are preserved** — existing cloud SDK
-  code runs unchanged after the upgrade.
-- The cloud API surface (`evaluations`, `galileo`, `cots`, `agentic`,
-  `frameworks`, `batch_jobs`, …) is untouched.
+The CLI command names (`trustmodel`, `trustmodel-mcp`) are unchanged; only the Python import module changed.
 
-## The one behavior change
+## The hosted cloud SDK is a SEPARATE package
 
-`auto_init` is now a dispatcher:
+`TrustModelClient` and the hosted API surface (`evaluations`, `galileo`, `cots`, `agentic`,
+`frameworks`, `guardrails.decide()`, `batch_jobs`, …) live in the **separate** official SDK:
 
-- `auto_init(api_key="tm-...", agent_id=..., domain=..., frameworks=[...])`
-  → routes to the OTel trace-forwarder, **exactly as in v2.x** (also reachable
-  explicitly at `trustmodel.telemetry.auto_init`).
-- `auto_init()` with no enterprise args → enables local inline scoring + optional
-  OpenTelemetry.
+```bash
+pip install trustmodel        # the hosted SDK
+```
+```python
+from trustmodel import TrustModelClient      # cloud SDK — NOT this repo
+```
 
-Existing `auto_init(...)` calls that pass `agent_id`/`domain`/`frameworks` are
-unaffected.
+Installing `trustmodel-local` does **not** give you `TrustModelClient`, and installing `trustmodel`
+does **not** give you the local `evaluate`/`monitor`/`govern`/`Guardrail` engine. Pick the one you need
+(or install both — they no longer collide).
 
 ## Requirements
 
 - The MCP server (`[mcp]` extra) requires **Python 3.10+** (the `mcp` SDK does).
-- The rest of the package still supports **Python 3.7+**.
+- The rest of the package supports **Python 3.9+**.
 
 ## Troubleshooting
 
-- `ImportError: cannot import name 'evaluate' from 'trustmodel'` → you're still
-  on v2.x. Run `pip install -U trustmodel`.
+- `ModuleNotFoundError: No module named 'trustmodel_local'` → `pip install -U trustmodel-local`.
+- `ImportError: cannot import name 'evaluate' from 'trustmodel'` → you're importing the local engine
+  from the wrong package. Use `from trustmodel_local import evaluate`.
 - `ModuleNotFoundError: No module named 'mcp'` when running `trustmodel-mcp` →
-  install the extra: `pip install -U "trustmodel[mcp]"` (and use Python ≥ 3.10).
-- `trustmodel: command not found` → pip installed the CLI to a per-user `bin/`
-  not on your `PATH`; run `python -m trustmodel ...` or add
-  `"$(python -m site --user-base)/bin"` to `PATH`.
+  `pip install -U "trustmodel-local[mcp]"` (and use Python ≥ 3.10).
 
 ## License note
 
-The engine (eval/monitor/govern + policy packs) is MIT. `TrustModelClient` is
-the proprietary cloud client. Both ship in the one wheel — see [LICENSE](LICENSE).
+The engine (eval/monitor/govern + policy packs) is MIT — see [LICENSE](LICENSE). The hosted
+`trustmodel` cloud SDK is a separate, proprietary package.
