+++ src/trustmodel_local/__main__.py
